chore(project): inline roots module
I’ve been wanting to do this for so long omg.
